+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## Unreleased
+
+* Adds the from_content_id builder method to LinkedContentItem
+
 ## 0.1.0
 
 * When parsing publishing api responses, allow both links and details hashes to